Generalize implementations of properties and modification events so that they can be mixed in to arbitrary classes.  Simplify property implementation to use an alist.  This saves 4 words of memory, which is important for message objects.
Create generalized resources, which are the referents of URLs.  Use
these as the base for folders, and create a new container type.

Change CREATE-FOLDER, DELETE-FOLDER, RENAME-FOLDER, and APPEND-MESSAGE
to signal appropriate events to the container of the folders being
manipulated.  This will allow folder browsers to automatically update
themselves as their contents are changed.

Add hooks to folder and container prompts, which allow the folder
browser to use some of the standard server commands.  The browser
can supply the appropriate URL string without prompting.

Don't let M-x imail-copy-folder copy from a folder to itself.