Revert earlier change to use RECORD-KEYWORD-CONSTRUCTOR, because it doesn't handle default values right.
Actually, I don't think it can do so, because DEFINE-STRUCTURE has
default expressions rather than default values; the expressions are
intended to be evaluated within the context of the constructor.  I
think this is a design flaw, but I'm not yet sure whether it is OK to
fix the design.