Change DEFINE-STRUCTURE to generate type descriptors for all structures, including untagged ones.  This will simplify some operations that need access to the type descriptor.
The default name to which the type descriptor is bound has been
changed to RTD:foo where "foo" is the structure's root name.

For the runtime cold load, allow TYPE-DESCRIPTOR option to accept #F
as an argument so that the structures defined in "packag.scm" don't
try to build a type descriptor.  This is important because this file
is loaded prior to the type-descriptor infrastructure.  A consequence
of this change is that the TYPE-DESCRIPTOR option no longer implies
tagging.  This is independently specified by the NAMED option, and
these two options are permitted to be used together.

Add TAG and OFFSET fields to the runtime type descriptor for
non-record structures.  In the next revision, this will allow building
more efficient constructors.