For backward compatibility, if new `call-with-output-string' is not defined, try using the old `with-string-output-port' name instead.
In short, 7.7 defines a backward-compatible package alias but 7.6 does
not define a forward-compatible anticipatory alias (obviously).

This is needed when, say, building 7.7 CVS sources using release 7.6
installed binaries as a bootstrap Scheme:  the new "makegen.scm" uses
the new name but old bootstrap "runtime"s may define only the old name.

This patch, coupled with the ${SCHEME_LARGE} and ${SCHEME_COMPILER}
generalizations of the `*.sh' build scripts, make it possible to call
"src/Setup.sh" to transition a CVS source directory to a "distribution"
state without first having the CVS Scheme installed to bootstrap.  This
consequently avoids a particularly subtle chicken-and-egg puzzle.

A forthcoming update to "src/README.txt" spells out the details.