Tighten XML-CHAR-DATA? and CANONICALIZE-CHAR-DATA so that they refuse characters or entity references not included in the XML-CHAR alphabet.
This may have a small performance cost when constructing XML trees,
and it will have to change for XML 1.1, which relaxes the restriction
on XML character data.  It could have been done in the XML writer,
since the parser already checks this, but this way offers guarantees
about the validity of any XML data in Scheme, and these guarantees are
probably good things.