Added `port-position' and `set-port-position!' operations on file I/O ports.  The `port-position' procedure returns the offset, in bytes, from the beginning of the file.  The `set-port-position!' procedure sets the position.
Both operations work only on ports that are opened with binary
normalizers or denormalizers, i.e. ones that make no transformation.
We should consider lifting this restriction, but this is a useful
addition even with it.

When used with output ports, the port passed to these procedures must
also be an input port.

The `port-position' procedure works even after a character has been
"unread", even for ports with non-single-byte character encodings.