Use a modification count cache to verify and commit memoized folder orders, instead of blocking interrupts, so that MEMOIZE-FOLDER-ORDER may now be interrupted.
Memoize the folder order immediately before it is assigned to a folder
in SET-FOLDER-ORDER!.  This means that it can no longer be delayed
until actually needed, but also that if it is interrupted it won't
actually set the folder's order and thereby wedge IMAIL until the
messages are sorted.

Allow message indices to fall outside the range of folder orders, and
leave them as they are, in case the folder order has not been updated
yet to reflect them.