Don't record the folder's cached modification count in folder orders; simply check that it has not changed since we built the tree.
New procedure MESSAGE-ORDER-KEY computes a message's key into the
folder order tree, if the message's folder is ordered.  This must be
computed before a message is actually expunged, because the key may
require information that is destroyed when expunging the message.
EXPUNGE modification events on folders now accept an extra parameter,
the key of the message being expunged.