New Edwin variable IMAIL-INLINE-MIME-TEXT-LIMIT, to prevent gargantuan text parts that satisfy all the other rules for in-line display from screwing Edwin.
This affects the main body of the message, too, however.  We could
work around this easily for unipart messges, although anything else
would require a little more effort.  In any case, although it is a
minor annoyance to have to type `C-c C-t C-e' to show the body of a
very long message, it is a much more major annoyance for Edwin to lose
spectacularly for a message too large to fit in its heap.