Rework mmap_heap_malloc so that it checks /proc/<pid>/maps only on Linux, and on other systems just goes ahead and tries a fixed map.
Previously, mmap_heap_malloc would try a fixed map, and if it failed,
try an unfixed map; then it was changed so that it would try a fixed
map only if it knew there was a good address.  I think that this only
affects Linux, so on other systems it will again try a fixed map first,
and an unfixed map only if that fails.