Re-enable open-coding of FLONUM-ASIN and FLONUM-ACOS on i386, and implement FLONUM-EXP using F2XM1 for the fractional part and FSCALE for the integral part of the argument.  This about trebles the speed of these primitives in compiled code, without giving answers substantially different from what the microcode computes.
When open-coding of FLONUM-EXP was disabled in 1992, a comment was
added to the effect that some i387 implementations had bugs in their
F2XM1 instructions.  I imagine that these bugs have been fixed in more
recent hardware, and that few users care about seventeen-year-old i387
units.  If this is a problem we can always disable it again.