Make X11 modules dynamically loadable.
Make the modules depend on `scheme' only for Mac OS X, and let that
be an option to configure, so that users can just build modules
against an already installed Scheme.

This is a step toward making every feature that currently involves a
compile-time decision instead be dynamically loadable, and compilable
separately from the microcode executable.  All that is left is
termcap support.