Make makegen/m4.sh report m4 failure by returning 1.
Previously, the return code of m4 was totally ignored, because it was
invoked only in a pipeline not at the end.  Because of brain damage
in shell script, we have to use a temporary file to let the subshell
in which we run m4 report the failure to the script so that the
script can know to return 1.