Be slightly more consistent about references to MIT_SCHEME_EXE.
It is to be interpreted as a pathname, provided that no character in
it be a make meta-character, or a shell meta-character that has
meaning within double-quotes.  It shouldn't have such constraints,
but Unix sucks.  To set the library path, use MITSCHEME_LIBRARY_PATH
rather than sticking a `--library' argument in MIT_SCHEME_EXE.