Fix i386 code generated for flonum rounding primitives.
The default has not been round-to-even since primitives were added to
change it, but the generated code assumed it was.

Remember to run at least a two-stage build so that the runtime is
compiled with the new compiler.

Before:

(flo:with-rounding-mode 'TOWARD-ZERO (lambda () (flo:ceiling 1.2)))
;Value: 1.

After:

(flo:with-rounding-mode 'TOWARD-ZERO (lambda () (flo:ceiling 1.2)))
;Value: 2.