Add configure option `--with-termcap=LIB' to select termcap library.
LIB=no disables any termcap library (causing the use of the built-in
termcap emulation); LIB=yes, the default, makes configure choose
whatever termcap library looks best.

I wanted to go further and make the termcap primitives dynamically
loadable, so that the compile-time choice of termcap library can be
made just for some prtermcap.so module separately from the libraries
against which the microcode is linked, but uxtty.c uses termcap for
screen-clearing and window-sizing.

If ^L worked everywhere to clear the screen, and TIOCGWINSZ worked
everywhere to find the screen dimensions, then we could get rid of
those uses of termcap and move it off to a module.  But alas, that is
not the case.  E.g., ^L doesn't clear the screen in screen.