New procedure REFERENCE-BARRIER.
(REFERENCE-BARRIER <x>) guarantees that the garbage collector will
assume <x> to be live at least until the call to REFERENCE-BARRIER,
if control can ever reach it.

This is implemented as an alias for the identity procedure currently,
but a more efficient implementation not involving a general procedure
call is conceivable for the future.