Implement user-defined unparser methods for entities.
New DEFINE-STRUCTURE option PRINT-ENTITY-PROCEDURE is like
PRINT-PROCEDURE, except that the second argument to the procedure
is not a structure instance itself but an entity whose extra is a
structure instance.

New procedure SET-RECORD-TYPE-ENTITY-UNPARSER-METHOD! is like
SET-RECORD-TYPE-UNPARSER-METHOD! with the same difference.

New optional argument to MAKE-RECORD-TYPE specifies an entity
unparser method.

Existing code should be unaffected by the changes, including
existing compiled code that used DEFINE-STRUCTURE.