Simplify and fix X-DISPLAY-PROCESS-EVENTS primitive.
This primitive no longer supports options for blocking vs
non-blocking or selecting vs not selecting; these are now the
responsibility of the caller.

The cached event trick previously used the value of the
uninitialized variable `event', rather than the event that was
cached.  Now it correctly uses the event that was cached.