New operations on the two's-complement representation of integers.
These include the SRFI 33 operations, as well as some other useful
operations.

Although these are implemented as primitives with native definitions
for bignums, the primitives are not yet open-coded for the fixnum
case.  Eventually they should be open-coded, so that you don't need
to make a choice between safe code using the integer operations and
fast code using FIX:AND, FIX:LSH, &c.  Some operations are easy to
open-code for the fixnum case, such as all the bitwise operations.
Others are not so easy, such as SHIFT-LEFT, and it would be better to
open-code common aggregate operations such as EXTRACT-BIT-FIELD for
the fixnum case.  In any case, at least we now have names for the
safe operations.