Trying to get FFI test shim to link on Mac OS X.
Following the linking of microcode modules more closely, including
SHIM_LOADER in the prerequisites for the -shim.so, and the curiously
brief SHIM_LIBS for the LINK_SHIM commandline.  Both are from
microcode/makegen/Makefile.in.in analogs, like MODULE_LIBS.

I'm not sure that this helps the Mac OS X build at all... so I'm
leaving build-ffis (and its test) disabled.