Move interrupt branch from start to end of block, on i386.
This makes it a forward branch rather than a reverse branch, so that
the CPU's branch predictor will predict it not taken rather than taken
by default.

Reduces time to compile the whole system by about 20%.  This
performance increase will apply to any compiled code, independent of
what sort of computation it does.