Remove all trace of type codes COMBINATION-2, PCOMB0, etc.
This finishes what 3fc580e started.  The resulting system runs but
does not build with the released version 9.1.  That version puts old
combinations in .exts that cannot be fasloaded by this new version.
As the old combination types are no longer compilable nor even
interpretable, it seems the build process loads them but does not use
them.

The following commit will put just the gc handlers back, which is
sufficient to allow building directly from 9.1.