cref: Support (parent #f) packages.  Punt system-global-package.
The package-structure<? sort did not ensure that BOTH a package's
parent-in-name and parent-environment were created first.  An easier
sort ensures only that the parent-environment is created first.  The
parent-in-name was only needed to hang the "child" on a tree.  This
patch replaces the tree with a list: *packages*.  Thus a
parent-in-name is not needed and the only parent/child tree is the
environment tree.

To catch out old code searching for packages via package/children (as
in swank.scm) the system-global-package binding was removed, as was
package/child.  Package/add-child! seems popular and so is supported
(with only minor trouble, i.e. no APPENDing during the cold load).
Package creation by other means is... not supported.