First stab at making top-level makefile jobs-safe for native builds.
This enables coarse-grained parallelization of the build.  There's a
nontrivial amount of crap encoded in the makefile about how to do the
build, but that's better than the nontrivial amount of dependency
crap implied by the total ordering in etc/compile.scm, most of which
can go away once the liarc and svm builds are converted similarly.

This does not restructure the build; we still have various bootstrap
and staging botches going on which we need to get rid of.  Fixing
those nicely will require some changes to liar and sf so that we can
compile a fresh compiler into a subdirectory, say `boot', using the
host compiler, and then use that to compile the system.  Liarc
approximates that with the `boot-root' installation prefix, but that
kludge breaks incremental builds which means we never, ever try to
touch the liarc build because the turnaround time is too painful.