Remove logic to lock IMAIL's IMAP folder cache.
Keep WITH-FOLDER-LOCKED as an empty stub so we might use it later to
let the user control whether to use a cache at all.

There are no invariants that need to be preserved, so locking only
slows us down and reduces concurrency.  Two clients fetching the same
item may collide, but the collision is harmless -- unless the server
is bogus, they will be writing the same data.  Adding an entry to the
cache is already atomic (unless the system fails before the blocks
have hit the disk) through the use of a temporary directory.