Fix new build to cross from 9.1.1.
On a 9.1.1 host, the new build dies while compiling the compiler.

    ;  Generating SCode for file: "fggen/canon.scm" => "fggen/canon.bin"...
    ;MICROCODE-TYPE: Unknown name sequence

The usual remedies did not suffice.  In fact now the temporary code in
cref/make.scm is not needed.  The compiler is compiled with a cross
runtime that provides the necessary patch via its utabs.scm, where
access to type-aliases and returns-aliases is easy, though
insufficient to train the host.  9.1.1 still needs the training in
runtime.sf.

Unlike a cross compile, a cross runtime was not used by this build(?).
Use a new compile-cross-runtime target and --fasl make.com to make
tools/syntaxer.com.  Make syntax-cross-cref and syntax-cross-sf
dependent on syntax-cross-runtime for its runtime.pkd (pinching off a
little parallelism, but allowing them to proceed while the runtime
compiles).  Did not try to syntax anything but the new compiler with
the new star-parser.