ffi: Punt AUXDIR from shim-config.scm.
Procedures like compile-shim need to use the first directory in
MITSCHEME_LIBRARY_PATH (not the installation directory) so that the
FFI test can override it, getting the not-yet-installed configuration
from the build directory.

Symlink shim-config.scm to src/lib/ and generate it during `make all',
not `make install', so that it is available to the test FFI (during
`make check').

Fix FFI test NOT to use with-notification around `make', which cannot
use fresh-line.  Just write whole lines to notification-output-port.