Unfluidize (runtime compress) internals, e.g. root-nodes.
This is the beginning of the end for fluid-let.

Gather all of the fluid variables into a compression-state object.
Pass it along as the first argument to many procedures.  The result is
approx. 10% slower, doing a lot of type and range checking where the
original was skipping checks for reference-traps.  A file-wide
(declare (no-range-checks) (no-type-checks)) got that 10% back PLUS 15%.